+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `CadenzaClient#fetchData`
+- `CadenzaClient#downloadData`
+
 ## 1.0.0 - 2023-09-07
 
 ## 0.0.1 - 2023-09-07
+
 ### Added
+
 - `CadenzaClient#show`
 - `CadenzaClient#showMap`
 - `CadenzaClient#editGeometry`